requirements.txt
Django>=2.2.0,<2.3.3
djangorestframework>=3.9.0,<3.10.0
